Refactoring(lab 5)

+++ lab1/Lab1/src/main/java/Classes/Engine.java
@@ -9,7 +9,7 @@ public class Engine {
      * @param capacity
      * @param manufacturer
      */
-    public Engine(int capacity, String manufacturer) {
+    public Engine(double capacity, String manufacturer) {
         Capacity = capacity;
         Manufacturer = manufacturer;
     }
